Fix: App init increment in login - increment only when user is trying to login with new zuid
This is happen in the scenario where - user logs out, closes app, opens app,
app init get called with new zuid, now login is called with the same new zuid.
In this case, user should be moved to old id, delete mixpanel user created with new zuid
and app init should increment. In other cases, this is not needed.